% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_polygon_list.R
\name{clip_polygon_list}
\alias{clip_polygon_list}
\title{Uses a polygon list to clip a given raster}
\usage{
clip_polygon_list(sample_polygons, sds_choose, file_save)
}
\arguments{
\item{sample_polygons}{should be a polygon list in SF language}

\item{sds_choose}{is a spat raster dataset that will be clipped for each polygon}

\item{file_save}{the location to store the clipped polygons}
}
\description{
Uses a polygon list to clip a given raster
}
\examples{
\dontrun{

file_save <- c('.')
sample_polygons <- read_sf('./shapefile.shp')
sds_to_run <- terra::sds(list.files('./stacked_metrics', pattern =- '00.tif$', full.names = T))

clip_polygon_list(sample_polygons, sds_choose = sds_to_run, file_save = file_save)
}

}
