% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_annum_sds.R
\name{mean_annum_sds}
\alias{mean_annum_sds}
\title{Calculates the mean and standard deviation of annual metrics from BAP SDS}
\usage{
mean_annum_sds(sds_choose, polygon)
}
\arguments{
\item{sds_choose}{this is a spatrasterdataset where each spat raster is stacked annually}

\item{polygon}{a polygon of some vector that is the same projection of the spat raster used to crop the sds}
}
\description{
Calculates the mean and standard deviation of annual metrics from BAP SDS
}
