% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_metrics_sds.R
\name{temporal_metrics_sds}
\alias{temporal_metrics_sds}
\title{Calculates a temporal values for a user defined time period and user defined function. Example function included in the documentation.}
\usage{
temporal_metrics_sds(
  sds_choose,
  time_length,
  last_year,
  eval_funct,
  polygon,
  disturbance_year,
  parent_folder
)
}
\arguments{
\item{sds_choose}{this is a spatrasterdataset where each spat raster is stacked annually}

\item{time_length}{what is the time length to calculate over ?}

\item{last_year}{last year of the dataset to calculate last metric to go by}

\item{eval_funct}{functions to evaluate for temporal metrics}

\item{polygon}{region to clip spat raster dataset by}

\item{disturbance_year}{optional argument to state what you would like to calculate by}

\item{parent_folder}{a folder that designates where to store the files}
}
\value{
metrics
}
\description{
Calculates a temporal values for a user defined time period and user defined function. Example function included in the documentation.
}
\examples{
\dontrun{
A an examples of the function summary input by the user
funSummary <- function(x){

  c(
    median = median(x,na.rm=TRUE),
    IQR = IQR(x,na.rm=TRUE),
   slope = foster::theilSen(x)
  )
}
}
}
