% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_metrics_sds.R
\name{temporal_sds}
\alias{temporal_sds}
\title{Calculates the mean and standard deviation of annual metrics from BAP SDS}
\usage{
temporal_sds(
  sds_choose,
  time_length,
  last_year,
  eval_funct,
  polygon,
  disturbance_year
)
}
\arguments{
\item{sds_choose}{this is a spatrasterdataset where each spat raster is stacked annually}

\item{time_length}{what is the time length to calculate over ?}

\item{last_year}{last year of the dataset to calculate last metric to go by}

\item{eval_funct}{functions to evaluate for temporal metrics}

\item{polygon}{region to clip spat raster dataset by}

\item{disturbance_year}{optional argument to state what you would like to calculate by}
}
\description{
Calculates the mean and standard deviation of annual metrics from BAP SDS
}
\examples{
A an examples of the function summary input by the user
funSummary <- function(x){

  c(
    median = median(x,na.rm=TRUE),
    IQR = IQR(x,na.rm=TRUE),
   slope = foster::theilSen(x)
  )
}
}
