% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_polygon_list.R
\name{clip_polygon_list}
\alias{clip_polygon_list}
\title{creates a polygon list to work with}
\usage{
clip_polygon_list(sample_polygons, r, return_rasters)
}
\arguments{
\item{sample_polygons}{should be a polygon list in SF language}

\item{return_rasters}{would yuo like to return the raster list?}

\item{raster}{is a raster that will be clipped for each polygon}
}
\description{
creates a polygon list to work with
}
